refactor: improve ui
- implement sidebar
- remove payment and receipt creating feat in all sales page
