+++ src/pages/NewOrder.jsx
@@ -1,21 +1,17 @@
 /* eslint-disable no-unused-vars */
-import { Button, Form, Grid } from "antd";
-import { useDispatch, useSelector } from "react-redux";
+import { Button, Form } from "antd";
+import { useDispatch } from "react-redux";
 import OrderDetailsTable from "../features/orders/OrderDetailsTable";
 import CreateOrderForm from "../features/orders/CreateOrderForm";
 import { clearOrder, clearOrderDetails } from "../features/orders/orderSlice";
 import SearchOrderDetail from "../features/orders/SearchOrderDetail";
 import SearchOrderCustomer from "../features/orders/SearchOrderCustomer";
-import CreateReceipt from "../features/receipts/CreateReceipt";
 import { useCreateOrder } from "../features/orders/hooks/useCreateOrder";
 
-const { useBreakpoint } = Grid;
 function NewOrder() {
   const dispatch = useDispatch();
   const [createOrderForm] = Form.useForm();
-  const customer = useSelector((state) => state.order.customer);
   const { createOrder, isCreating } = useCreateOrder();
-  const screens = useBreakpoint();
 
   function handleFinish(submittedOrder) {
     createOrder(submittedOrder, {
@@ -42,10 +38,7 @@ function NewOrder() {
         <OrderDetailsTable />
       </div>
       <div className="card flex min-h-[calc(100vh-64px-1.5rem*2)] basis-[30%] flex-col justify-between gap-8">
-        <div className="flex items-center justify-between">
-          <SearchOrderCustomer />
-          {screens.xl && <CreateReceipt customer={customer} />}
-        </div>
+        <SearchOrderCustomer />
         <div className="flex-1">
           <CreateOrderForm form={createOrderForm} onFinish={handleFinish} />
         </div>
