--- src/ui/Main.jsx
@@ -1,8 +1,0 @@
-import { Layout } from "antd";
-
-/* eslint-disable react/prop-types */
-function Main({ children }) {
-  return <Layout.Content>{children}</Layout.Content>;
-}
-
-export default Main;
